@@ -1,5 +1,5 @@
 [tool.black]
-target-version = ["py310"]
+target-version = ["py311"]
 
 [tool.coverage.html]
 title = "{{cookiecutter.project_name}} - Coverage"
@@ -21,28 +21,54 @@ omit = [
     "manage*.py",
     "pacts/*",
     "venv/*",
+    "*/pact_states.py",
 ]
 source = ["."]
 
-[tool.isort]
-atomic = true
-combine_as_imports = true
-default_section = "THIRDPARTY"
-known_first_party = [
-    "{{ cookiecutter.django_settings_dirname }}",
-]
-known_third_party = [
-    "django",
-]
-profile = "black"
-skip = [
-    ".vscode-server",
-]
-
 [tool.mypy]
 ignore_missing_imports = true
-python_version = "3.10"
+python_version = "3.11"
 
 [tool.pytest.ini_options]
 DJANGO_CONFIGURATION = "Testing"
 DJANGO_SETTINGS_MODULE = "{{ cookiecutter.django_settings_dirname }}.settings"
+
+[tool.ruff]
+extend-exclude = [
+    "__pycache__",
+    ".pytest_cache",
+    ".vscode*",
+    "*/migrations/*",
+]
+extend-ignore = [
+    "D203",
+    "D212",
+    "D213",
+    "D214",
+    "D215",
+    "D404",
+    "D405",
+    "D406",
+    "D407",
+    "D408",
+    "D409",
+    "D410",
+    "D411",
+    "D413",
+    "D415",
+    "D416",
+    "D417",
+]
+select = ["A", "B", "C", "D", "E", "F", "I", "W", "B9"]
+target-version = "py311"
+
+[tool.ruff.per-file-ignores]
+"{{ cookiecutter.django_settings_dirname }}/*sgi.py" = ["E402"]
+
+[tool.ruff.isort]
+known-first-party = [
+    "{{ cookiecutter.django_settings_dirname }}",
+]
+known-third-party = [
+    "django",
+]
